V-2.44.10 - CAD: projection-plane added to View Top, Front, Side - MAN: new function search-object (Ctrl-F) - bugfix 'move-points_on_construction-plane' - bugfix 'resize-window-smaller' - bugfix 'CentOS-nan-crash' - bugfix 'Debian-startup-hang-uname'

+++ src/ci/NC_Main.h
@@ -32,7 +32,7 @@ void   WC_EOF            ();
 int    WC_PP_open        (int mode);
 
 void   WC_setDisp3D      (int);
-int    NC_setRefsys      (int);
+int    NC_setRefsys      (long);
 void   NC_setGrafAtt     (int AttInd);
 void   NC_set_actPos     (Point2*);
 int    WC_set_obj_stat   (int stat);
